requirements.txt
pytest
ruff
mypy
pytest-cov
openai
pyaudio
python-dotenv
numpy
